[Build] Update debuild for Ubuntu Focal
Ubuntu Focal no longer contains a `python` package, so launchpad nightly
builds fail with a "dependency wait" error for Focal.

This updates the dependencies to default to `python3` package, but
allows for a fallback on a `python` package to maintain backwards
compatibility.

Also added the Qt5 SVG and (if available) the Qt5 Charts packages.