Allow for custom zk params dir
This adds two methods of defining a custom zk params directory:

1. Runtime option `-paramsdir`, which functions exactly like `-datadir`.
2. Configure/Compile option `--with-params-dir` as a way to hard-code
the zk params directory at compile time.

The second method was added because our CI windows tests made the
assumption that the test environment was windows, which it isn't, so it
was trying to reference a path that doesn't exist.