[Init] More robust early MN config checking
Our early MN config checking is rather limited, and actually incorrect
right now. It is completely valid to only listen on the network type
that the MN runs on (ie, a tor-only MN is not required to listen on
clearnet).

This re-works our early initialization sanity checks to both fix the
erroneous logic, but also expand it to detect missing config flags as
well.